allow writing to content store directly
The primary benefit of this is that you can import OCI archives directly
into Buildkit and then fetch them using OCIStore attachable. Without
this change you have to unpack to a local OCI store and then stream it
in via the attachable, which is twice the amount of work and requires
writing to the client-side filesystem.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>